+++ src/Integrations/Plugins/AbstractComponent.php
@@ -4,6 +4,7 @@ namespace JtlWooCommerceConnector\Integrations\Plugins;
 
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
 
 /**
  * Class AbstractComponent
@@ -20,6 +21,15 @@ abstract class AbstractComponent implements ComponentInterface, LoggerAwareInter
      * @var PluginInterface
      */
     protected $plugin;
+
+
+    /**
+     * @param LoggerInterface|null $logger
+     */
+    public function __construct(?LoggerInterface $logger = null)
+    {
+        $this->logger = $logger ?? new NullLogger();
+    }
 
     /**
      * @param LoggerInterface $logger